+++ ci/playbooks/containers/podman_containers.yml
@@ -1,0 +1,9 @@
+---
+- hosts: all
+  gather_facts: true
+  tasks:
+    - include_role:
+        name: podman_containers
+      vars:
+        idem_image: idempotency_test
+        ansible_python_interpreter: "{{ _ansible_python_interpreter }}"
