+++ tests/sanity/ignore-2.10.txt
@@ -1,5 +1,4 @@
 plugins/modules/podman_image.py validate-modules:mutually_exclusive-unknown
-plugins/modules/podman_container.py validate-modules:invalid-ansiblemodule-schema
 plugins/modules/podman_container.py validate-modules:parameter-list-no-elements
 plugins/modules/podman_container.py import-2.6!skip
 plugins/modules/podman_container.py compile-2.6!skip
@@ -9,6 +8,14 @@ plugins/modules/podman_container.py import-3.6!skip
 plugins/modules/podman_container.py import-3.7!skip
 plugins/modules/podman_container.py import-3.8!skip
 plugins/modules/podman_container.py import-3.9!skip
+plugins/modules/podman_containers.py import-2.6!skip
+plugins/modules/podman_containers.py compile-2.6!skip
+plugins/modules/podman_containers.py import-2.7!skip
+plugins/modules/podman_containers.py import-3.5!skip
+plugins/modules/podman_containers.py import-3.6!skip
+plugins/modules/podman_containers.py import-3.7!skip
+plugins/modules/podman_containers.py import-3.8!skip
+plugins/modules/podman_containers.py import-3.9!skip
 plugins/modules/podman_pod.py import-2.6!skip
 plugins/modules/podman_pod.py compile-2.6!skip
 plugins/modules/podman_pod.py import-2.7!skip
@@ -25,6 +32,15 @@ plugins/modules/podman_volume.py import-3.6!skip
 plugins/modules/podman_volume.py import-3.7!skip
 plugins/modules/podman_volume.py import-3.8!skip
 plugins/modules/podman_volume.py import-3.9!skip
+plugins/module_utils/podman/podman_container_lib.py import-2.6!skip
+plugins/module_utils/podman/podman_container_lib.py compile-2.6!skip
+plugins/module_utils/podman/podman_container_lib.py import-2.7!skip
+plugins/module_utils/podman/podman_container_lib.py import-3.5!skip
+plugins/module_utils/podman/podman_container_lib.py import-3.6!skip
+plugins/module_utils/podman/podman_container_lib.py import-3.7!skip
+plugins/module_utils/podman/podman_container_lib.py import-3.8!skip
 tests/integration/targets/connection_buildah/runme.sh shellcheck:SC2086
 tests/integration/targets/connection_podman/runme.sh shellcheck:SC2086
 plugins/module_utils/podman/common.py shebang!skip
+plugins/doc_fragments/podman_containers_doc.py future-import-boilerplate!skip
+plugins/module_utils/podman/podman_container_lib.py future-import-boilerplate!skip
