+++ tests/sanity/ignore-2.9.txt
@@ -5,6 +5,13 @@ plugins/modules/podman_container.py import-3.5!skip
 plugins/modules/podman_container.py import-3.6!skip
 plugins/modules/podman_container.py import-3.7!skip
 plugins/modules/podman_container.py import-3.8!skip
+plugins/modules/podman_containers.py import-2.6!skip
+plugins/modules/podman_containers.py compile-2.6!skip
+plugins/modules/podman_containers.py import-2.7!skip
+plugins/modules/podman_containers.py import-3.5!skip
+plugins/modules/podman_containers.py import-3.6!skip
+plugins/modules/podman_containers.py import-3.7!skip
+plugins/modules/podman_containers.py import-3.8!skip
 plugins/modules/podman_pod.py import-2.6!skip
 plugins/modules/podman_pod.py compile-2.6!skip
 plugins/modules/podman_pod.py import-2.7!skip
@@ -19,6 +26,15 @@ plugins/modules/podman_volume.py import-3.5!skip
 plugins/modules/podman_volume.py import-3.6!skip
 plugins/modules/podman_volume.py import-3.7!skip
 plugins/modules/podman_volume.py import-3.8!skip
+plugins/module_utils/podman/podman_container_lib.py import-2.6!skip
+plugins/module_utils/podman/podman_container_lib.py compile-2.6!skip
+plugins/module_utils/podman/podman_container_lib.py import-2.7!skip
+plugins/module_utils/podman/podman_container_lib.py import-3.5!skip
+plugins/module_utils/podman/podman_container_lib.py import-3.6!skip
+plugins/module_utils/podman/podman_container_lib.py import-3.7!skip
+plugins/module_utils/podman/podman_container_lib.py import-3.8!skip
 tests/integration/targets/connection_buildah/runme.sh shellcheck:SC2086
 tests/integration/targets/connection_podman/runme.sh shellcheck:SC2086
 plugins/module_utils/podman/common.py shebang!skip
+plugins/doc_fragments/podman_containers_doc.py future-import-boilerplate!skip
+plugins/module_utils/podman/podman_container_lib.py future-import-boilerplate!skip
